`CMAKE_SOURCE_DIR`/`PROJECT_DIR` -> `CMAKE_CURRENT_LIST_DIR`
In commit 9b90882146, initial support was added for building with the
newly introduced CMake support provided by the ESP-IDF.

Specifically, the commit message states:

> This commit adds support for building the esp32 port with CMake, and
in particular, it builds MicroPython as a component within the ESP-IDF.
Using CMake and the ESP-IDF build infrastructure makes it much easier to
maintain the port, especially with the various new ESP32 MCUs and their
required toolchains.

`PROJECT_DIR` is a variable populated by the ESP-IDF specifically and is
not stable when used with "[Pure CMake components][1]" as documented in
the ESP-IDF. It is intended to be used in the scope of the parent of the
current file (the "project") as opposed to the current file ("the
component"). Crossing into the parent scope like this works solely when
the "project" is MicroPython, but not when used as a component by other
ESP-IDF projects.

Analyzing this file, the intention is to reference the "Project" which
in the example is the parent directory. Within the [CMake variables][2]
documentation, there is one specifically defined for referencing the
directory for the CMake listfile currently being processed:
`CMAKE_CURRENT_LIST_DIR`.

After making the change from `PROJECT_DIR` to `CMAKE_CURRENT_LIST_DIR`,
the reach into the parent scope defined by the ESP-IDF and the resulting
CMake interface violation is removed.

Similar to the component definition, the project `CMakeLists.txt` uses
the variable `CMAKE_SOURCE_DIR` which CMake defines as "The path to the
top level of the source tree."  This commit changes the variable to
`CMAKE_CURRENT_LIST_DIR` for the reasons cited above.

[1]: https://docs.espressif.com/projects/esp-idf/en/latest/esp32s2/api-guides/build-system.html#writing-pure-cmake-components
[2]: https://cmake.org/cmake/help/latest/manual/cmake-variables.7.html
[3]: https://cmake.org/cmake/help/latest/variable/CMAKE_CURRENT_LIST_DIR.html